reduce lag time when loading AI suggestion
